- Added rules to delete professor - Fixed bugs in rules of delete department

+++ src/main/java/com/example/myapp/dao/impl/ProfessorDAOImpl.java
@@ -6,7 +6,12 @@ import java.util.ArrayList;
 
 import com.example.myapp.dao.ProfessorDAO;
 import com.example.myapp.factory.DatabaseConnection;
+import com.example.myapp.model.Department;
+import com.example.myapp.model.Participates;
 import com.example.myapp.model.Professor;
+import com.example.myapp.model.Project;
+import com.example.myapp.model.Student;
+import com.example.myapp.model.Works;
 import com.mongodb.BasicDBObject;
 import com.mongodb.client.FindIterable;
 import com.mongodb.client.MongoCollection;
@@ -15,9 +20,17 @@ import com.mongodb.client.MongoCursor;
 public class ProfessorDAOImpl implements ProfessorDAO{
 
 	private MongoCollection<Professor> collection;
+	private MongoCollection<Participates> participatesCollection;
+	private MongoCollection<Project> projectCollection;
+	private MongoCollection<Department> departmentCollection;
+	private MongoCollection<Works> worksCollection;
 	
 	public ProfessorDAOImpl() {
 		collection = DatabaseConnection.getConnection().getCollection("professor", Professor.class);
+		participatesCollection = DatabaseConnection.getConnection().getCollection("participa", Participates.class);
+		projectCollection = DatabaseConnection.getConnection().getCollection("projeto", Project.class);
+		departmentCollection = DatabaseConnection.getConnection().getCollection("departamento", Department.class);
+		worksCollection = DatabaseConnection.getConnection().getCollection("trabalha", Works.class);
 	}
 
 	private boolean exists(Professor professor) {
@@ -58,8 +71,20 @@ public class ProfessorDAOImpl implements ProfessorDAO{
 	}
 
 	@Override
-	public void delete(Long id) {
+	public boolean delete(Long id) {
+		
+		Participates participates = participatesCollection.find(eq("professor.reg_number", id)).first();
+		Project project = projectCollection.find(eq("profLeader.reg_number", id)).first();
+		Department department = departmentCollection.find(eq("profLeader.reg_number", id)).first();
+		
+		if(participates != null || project != null || department != null) {
+			return false;
+		}
+		
+		worksCollection.deleteMany(eq("professor.reg_number", id));
 		collection.deleteOne(eq("reg_number", id));
+		
+		return true;
 	}
 
 	@Override
